# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Needon Twitter Bot API",
    "description": "Description",
    "version": "1.0.0"
  },
  "host": "localhost:3001",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "paths": {
    "/twitter/likes": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "twitterId",
            "in": "query",
            "description": "트위터 아이디 입력",
            "required": false,
            "type": "string"
          },
          {
            "name": "size",
            "in": "query",
            "description": "출력 갯수",
            "required": false,
            "type": "integer"
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "url",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "로그인 API",
        "parameters": [
          {
            "name": "username",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "formData",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "로그인 성공",
            "schema": {
              "type": "object",
              "properties": {
                "token": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "로그인 실패"
          }
        }
      }
    },
    "/auth/protected": {
      "get": {
        "summary": "인증이 필요한 API",
        "security": [
          {
            "Bearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "인증 성공",
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "인증 실패"
          }
        }
      }
    }
  }
}